+++ src/almacenamiento/index.ts
@@ -1,0 +1,5 @@
+export * from './tipos'
+export * from './claves'
+export * from './adaptadorMemoria'
+export * from './adaptadorCapacitor'
+export * from './repositorio'